refactor(frontend): unified game module pattern + shared HTTP client (#153)
Collapses ~250 lines of duplicated transport logic across 5 client files
into one shared factory, and reorganizes per-game code into
frontend/src/game/<name>/{api,types}.ts. Establishes the directory
pattern that offline engine ports (#156, #157, #158, #159) will drop
into.

Changes:

- frontend/src/api/* → deleted. Each client moves to:
  - frontend/src/game/yacht/{api,types}.ts   (from api/client.ts)
  - frontend/src/game/cascade/{api,types}.ts
  - frontend/src/game/blackjack/{api,types}.ts
  - frontend/src/game/ludo/{api,types}.ts
  - frontend/src/game/twenty48/{api,types}.ts

- Each api.ts is now ~20 lines — thin wrapper over createGameClient()
  from _shared/httpClient.ts (introduced in PR #160). The per-game
  Sentry apiTag is preserved for observability.

- getOrCreateSessionId() extracted from api/client.ts to
  _shared/session.ts. Removes the awkward "every client imports from
  the Yacht client" reverse dependency.

- BASE_URL configuration tests collapsed from 4 duplicated suites
  (3 tests × 4 client files = 12 tests) into one shared
  _shared/__tests__/httpClient.test.ts. Same for error-shaping tests
  (statusText fallback, detail propagation). Per-game test files now
  focus only on endpoint-specific behavior (path, method, body).

- Yacht endpoints still use the legacy /game/* prefix — renaming to
  /yacht/* is a breaking API change tracked separately in #161.

No behavior changes. 413/413 Jest tests pass, 31/31 Playwright tests
pass, lint + prettier clean.
