+++ frontend/src/game/cascade/types.ts
@@ -1,0 +1,12 @@
+/**
+ * Cascade API response shapes.
+ */
+
+export interface ScoreEntry {
+  player_name: string;
+  score: number;
+}
+
+export interface LeaderboardResponse {
+  scores: ScoreEntry[];
+}
