+++ frontend/src/screens/Twenty48Screen.tsx
@@ -5,7 +5,7 @@ import { useTranslation } from "react-i18next";
 import { NativeStackNavigationProp } from "@react-navigation/native-stack";
 import { RootStackParamList } from "../../App";
 import { useTheme } from "../theme/ThemeContext";
-import { twenty48Api, Twenty48State } from "../api/twenty48Client";
+import { twenty48Api, Twenty48State } from "../game/twenty48/api";
 import Grid from "../components/twenty48/Grid";
 import ScoreBoard from "../components/twenty48/ScoreBoard";
 import GameOverlay from "../components/twenty48/GameOverlay";
