+++ frontend/src/screens/__tests__/HomeScreen.test.tsx
@@ -7,13 +7,13 @@ import { ThemeProvider } from "../../theme/ThemeContext";
 // ---------------------------------------------------------------------------
 // Mock the Yacht API client
 // ---------------------------------------------------------------------------
-jest.mock("../../api/client", () => ({
+jest.mock("../../game/yacht/api", () => ({
   api: {
     newGame: jest.fn(),
   },
 }));
 
-import { api } from "../../api/client";
+import { api } from "../../game/yacht/api";
 const mockApi = api as jest.Mocked<typeof api>;
 
 // ---------------------------------------------------------------------------
